+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Stanza now always selects the GPU with the most free memory.
 - The preloader can now be gracefully stopped by sending an interrupt signal to the Sparv process.
 - Added `HeaderAnnotations` and `HeaderAnnotationsAllSourceFiles` classes.
+- Added `korp.keep_undefined_annotations` setting, to include even undefined annotations in the Korp config.
 
 ### Changed